c64chars/requirements.txt
pillow  

